fix(intr): an annoying bug caused by integer overflow

+++ kernel/interrupt.c
@@ -185,7 +185,7 @@ void idt_init() {
   pic_init();
 
   // why uint32_t?
-  uint64_t idt_operand = ((sizeof(idt) - 1) | ((uint64_t)((uint32_t)idt << 16)));
+  uint64_t idt_operand = ((sizeof(idt) - 1) | ((uint64_t)idt << 16));
   asm volatile("lidt %0" : : "m" (idt_operand));
   put_str("idt_init done\n");
 }
